Move SearchResults to containers
